backend/pyproject.toml
[project]
name = "hiagent"
version = "0.1.0"
description = "HiAgent - AI Agent Framework"
requires-python = ">=3.12"
dependencies = [
    "anthropic",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "loguru",
    "redis",
    "httpx",
    "tavily-python",
    "boxlite",
    "pyyaml",
    "e2b-code-interpreter",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "api", "config"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
]
